requirements.txt
pydantic==2.9.2
fastapi==0.104.1
httpx==0.27.0
SQLAlchemy==2.0.25
PyJWT==2.9.0
aiohttp==3.9.1
aiocache==0.12.2
uvicorn==0.23.2
pytest==8.1.1
